Implement Single-file Bundler
This change implements the single-file bundler, according to this
[design document](https://github.com/dotnet/designs/blob/master/accepted/single-file/design.md)

The change adds a test-case to bundle the contents of a
self-contained console app, extract the contents, and run the app.

Other tests will be added once the host-changes to process the bundle
are implemented.

Issue: https://github.com/dotnet/coreclr/issues/20287