Single-file: Facilitate signing on mac
The single-file bundler adds additional contents (the apps's dependencies) at the end of the host-binary.
This violates the strict consistency checks performed by codesign tool in MAC.

This change adjusts the MachO headers to accomodate the bytes added by the bundler.

This Method is a utility to adjust the apphost MachO-header
to include the bytes added by the single-file bundler at the end of the file.

Fixes #7065